Fix orders loading in 3DS account info formatter
Using the same collection to load 3 different set of orders is wrong because Magento always use the firs loaded collection. Indeed with the previous code Magento was sending the count of orders in the last 6 month even for the last year orders count and the last 24 hours orders count.

Also note that the HiPay API limit the length of the last 24 hours orders count parameter to 3 characters length. So in a store where the same customer place more than 999 orders in 6 months he cannot pay anymore with the HiPay payment method.